requirements.txt
strawberry-graphql[fastapi]
fastapi
uvicorn[standard]
neo4j
python-dotenv 